Move ./utils/role.js -> ./api/roles.js
