+++ functions/venue.js
@@ -2,8 +2,9 @@ const admin = require("firebase-admin");
 const functions = require("firebase-functions");
 const { HttpsError } = require("firebase-functions/lib/providers/https");
 
+const { addAdmin, removeAdmin } = require("./src/api/roles");
+
 const { checkAuth } = require("./src/utils/assert");
-const { addAdmin, removeAdmin } = require("./src/utils/role");
 const { getVenueId, checkIfValidVenueId } = require("./src/utils/venue");
 
 const PLAYA_VENUE_ID = "jamonline";
